fill-column-indicator: add vertical bar face
This change is taken from
modus-themes (https://www.gnu.org/software/emacs/manual/html_node/modus-themes/Note-on-display_002dfill_002dcolumn_002dindicator_002dmode.html):
The display-fill-column-indicator-mode uses a typographic character to
draw its line. This has the downside of creating a dashed line. The
dashes are further apart depending on how tall the font’s glyph height
is and what integer the line-spacing is set to.

At the theme level we eliminate this effect by making the character
one pixel tall: the line is contiguous.